% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{ashake_dependency}
\alias{ashake_dependency}
\title{Dependency}
\usage{
ashake_dependency(cdn = FALSE)
}
\arguments{
\item{cdn}{Whether to use local or CDN files.}
}
\description{
Add ashake dependency.
}
\examples{
library(aframer)

browse_aframe(
  a_scene(
    a_dependency(),
    ashake_dependency(cdn = TRUE),
    a_entity(
      shake2show = NA,
      visible = "false",
      a_box(
        position = xyz_aframe(0, 1, 3),
        color = "black"
      )
    ),
    a_entity(
      camera = NA,
      `look-controls` = NA
    )
  )
)

}
